[Android] Catch-up with iOS for UI elements.
This covers quasi all elements found in the iOS app.  There is no functionality to talk to the API yet, and no UI elements for showing API results are included either.

I took the liberty to make some UI changes - like the HeaderView bar, to better align with Jetpack Compose's opinions and facilities - a top app bar in this case.

This also uses Jetpack Compose Navigation, though in hindsight it might be overkill.  That required us to use view models to persist basic state such as slider values (which I assume is a desire, though I noticed iOS doesn't implement it itself; filed https://github.com/nicbov/EatRight/issues/6).  Some solution for that would've been needed to support screen rotations too (but maybe `rememberSaveable` does the trick for that, and seems simpler?).

I toyed with the Material Theme Builder (https://material-foundation.github.io/material-theme-builder/) to come up with a fancy palette based on the two brand colors we have so far.  Despite checking 'Color match' the generated primary color remained much darker than the one inputted.  I see others complain about this (e.g., https://www.reddit.com/r/MaterialDesign/comments/15v1lsa/how_do_i_set_up_the_color_theme_for_material/).  I reverted to using the at-runtime computed palette, and postponing dark mode customizations.

I'm not super clear when a passed-on `modifier` should be used in a composable's implementation.  The docs say to apply it to the first or top-level UI, so I am consistently doing that now.  But it doesn't state if it should be used beyond that.  It shouldn't matter much since these composables are not really reusable anyhow.

There is a lot of styling spread out in the composable .kt files, things like padding, spacers, image sizes, that are probably not managed in a way that maintains consistency.  That is left as a future improvement.