Use `rememberCoroutineScope` rather than `viewModelScope` for the login coroutine.
The former is preferred when the coroutine is scoped to a composable, and the latter is used for longer / wider app scope.  Here it's fine for the coroutine to share the shorter lifecycle of the login screen.

Note that `LaunchedEffect` is an alternative for use in a composable scope, but it is a composable itself and not usable in e.g. a click event handler (as is the case here).  See https://stackoverflow.com/questions/66474049/using-remembercoroutinescope-vs-launchedeffect.