[Android] Fetch recipe HTTP client lifecycle: Survive recompositions.
While the client doesn't have state we care about across requests, the ktor doc says this isn't a light-weight object, so avoid creating one for each request, as well as recreating it during recompositions.  It seems light-weight enough not to care about the recreation cost across reconfigurations (screen rotations, which `rememberSaveable` would help with), navigation, etc, so a simple `remember` seems sufficient here.

There might be other reasons to hoist this up in the future though.  This might all better belong in a different file to start with. KISS for now.