Added SVG to mealdetailview to provide easy summary of taste info.
 Added Chart.py to handle the charting of the svg and save it as bytes
in mem. Created /taste_chart route to main.py that handles
recieving/returning the svg data. Made changes in MealDetailView and
RecipeService to request the SVG and display it. Currently, the frontend
takes the old taste_info supplied through the recipe/recipe_id endpoint,
then makes a new post request with that info to the /taste_chart
endpoint.