[Android] Meal detail screen.
This introduces a new navigation destination for a meal / recipe.  For the first time we have a route that includes a parameter, in this case the recipe ID.  This seems to be a good fit, except for the fact that the backend currently doesn't return the recipe title in its recipe details API, requiring us to cache this (including the title as a parameter in the route seems possible but not a good fit to me).

This introduces the use of AndroidSVG for rendering the fetched SVG of the taste info.  AndroidSVG is based on the old "Views" framework, but it seems to work fine and is isolated.

The IOS app uses a "DisclosureGroup" to expand on nutrition info etc.  There's no such standard widget in Jetpack compose, so we roll out our own.

The IOS app shows meal details within the header view, but that doesn't seem desired.  We implement a more full screen experience here, with just the logo and back-button, similar to the profile / settings pages.

As for the existing fetch code, work remains to handle errors and to separate this business logic from the UI.  This is left for later.

This includes a server-side change to ensure the pricing Amount field is always numeric - it is otherwise sometimes string sometimes number, which causes deserialization errors (unlike in IOS where it can just work with string).