Add an android container to the environment.
This is a convenient way to run the emulator in Linux.  On MacOS, it complicates things - or even makes it impossible, given the nested virtualization (the android container uses /dev/kvm - the linux VM on MacOS doesn't have that?).  We'll want to make this bit conditional, which appears to be possible in compose using 'profiles', if we have uses for compose in MacOS.