Build using host networking.
I don't know if this is a new docker default, or the default mode is broken, but just for building, without specifying this, there is no internet access on manon.  Luckily, there seems to be a docker compose way to configure the network mode just for building, without affecting the real containers, since we want those to use their own network.