[Android] Switch emalator to the aosp_atd system image.
This is a lighter image that google's, which doesn't have google's services.  We have no deps on those services as of now, and are running into flaky emulator runs (lack of CPU?).  We're hoping the lighter image will help with those.

API 31 was tried as that's the min declared for our app, in the hope it would additionally help in terms of light-weightedness.  However the emulator bailed out.  Errors suggested this could be due to Vulkan failures.  When trying API 34, the crash no longer appeared, but there were still Vulkan errors - so those might not have been the cause.  Still, a fix for the Vulkan error was researched and included - looks like we just needed some additional libraries in the image.  Vulkan might not ever be relevant, and it's possible we are better off in the long term without its support.  But it seems best to avoid the error as it might confuse future travelers.