Switch back to google APIs Android image.
This effectively rolls back c404aae (minus orthogonal misc. changes we'd like to keep).

Why?  I noticed the captured videos are all just a grey screen in the last builds.  I find https://issuetracker.google.com/issues/259111479?pli=1 which suggests this can happen with swiftshader.  I remember the Vulkan issues with the aosp_atd image and seeing it is using (reverting?) to swiftshader -- this may be tricky to fix.

I tried the google APIs image and can indeed record a valid screen again.  We have doubled the RAM on the Jenkins server since c404aae, so no longer have the same motivation to use a lighter image.  We may need the APIs in the future as we integrate with more services during app development.  So, let's give this image another try.