[CI] Suppress confusing build duration suffix in Discord notifications.
E.g.:

```
Build ID: 598
Build Status: SUCCESS
Build Duration: 5 min 51 sec and counting
```

"and counting" is technically correct since the build is still running when we send this notification.  But it is confusing to a reader on Discord, since it's supposed to be the notification of the final state of the build.

How?  Picked approach from https://github.com/jenkinsci/slack-plugin/issues/327.