[CI] Attempt at fixing the CI pipeline with the new backend.
There is no more Fake API container, but we have a mysql one.  The database needs to be populated on startup.  This fails if the container was previously started.

Renamed the docker compose file to use the standard naming.

Including an idea file -- still not sure if checking these in is the right strategy.  Anyhow, unrelated to the rest of the commit for sure.